requirements.txt
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
click==8.1.3
greenlet==1.1.2
h11==0.13.0
numpy==1.23.1
pandas==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
uvicorn==0.18.2
